+++ config/routes.rb
@@ -20,7 +20,7 @@ Rails.application.routes.draw do
 
   devise_scope :user do
     root to: "auth/sessions#new"
-    put "disable_user/:id", to: "auth/registrations#disable", as: :disable_user
+    put "toggle_enabled/:id", to: "auth/registrations#toggle_enabled", as: :toggle_enabled
   end
 
   namespace :v2, module: "api/v2", defaults: { format: :json } do